Adapt to refactoring of GModel source model classes & log4j
  * Adaptations for https://github.com/eclipse-glsp/glsp-server/pull/165
  * TPD update, see https://github.com/eclipse-glsp/glsp-server/pull/163
