+++ server/example/org.eclipse.glsp.ide.workflow.editor/src/org/eclipse/glsp/ide/workflow/editor/WorkflowGLSPEclipseModule.java
@@ -29,9 +29,9 @@ import org.eclipse.glsp.ide.editor.actions.handlers.IdeSetMarkersActionHandler;
 import org.eclipse.glsp.ide.editor.actions.handlers.InitializeCanvasBoundsActionHandler;
 import org.eclipse.glsp.ide.editor.actions.handlers.SetClipboardDataActionHandler;
 import org.eclipse.glsp.ide.editor.di.IdeActionDispatcher;
+import org.eclipse.glsp.ide.editor.gmodel.operations.IdeGModelPasteOperationHandler;
 import org.eclipse.glsp.ide.editor.initialization.DefaultModelInitializationConstraint;
 import org.eclipse.glsp.ide.editor.initialization.ModelInitializationConstraint;
-import org.eclipse.glsp.ide.editor.operations.handlers.IdePasteOperationHandler;
 import org.eclipse.glsp.server.actions.Action;
 import org.eclipse.glsp.server.actions.ActionDispatcher;
 import org.eclipse.glsp.server.actions.ActionHandler;
@@ -40,8 +40,8 @@ import org.eclipse.glsp.server.actions.ServerStatusAction;
 import org.eclipse.glsp.server.actions.SetDirtyStateAction;
 import org.eclipse.glsp.server.di.MultiBinding;
 import org.eclipse.glsp.server.features.navigation.NavigateToExternalTargetAction;
+import org.eclipse.glsp.server.gmodel.GModelPasteOperationHandler;
 import org.eclipse.glsp.server.operations.OperationHandler;
-import org.eclipse.glsp.server.operations.gmodel.PasteOperationHandler;
 
 import com.google.inject.Scopes;
 
@@ -72,8 +72,8 @@ class WorkflowGLSPEclipseModule extends WorkflowDiagramModule {
    @Override
    protected void configureOperationHandlers(final MultiBinding<OperationHandler> bindings) {
       super.configureOperationHandlers(bindings);
-      bindings.remove(PasteOperationHandler.class);
-      bindings.add(IdePasteOperationHandler.class);
+      bindings.remove(GModelPasteOperationHandler.class);
+      bindings.add(IdeGModelPasteOperationHandler.class);
    }
 
    @Override
