+++ server/plugins/org.eclipse.glsp.ide.editor/META-INF/MANIFEST.MF
@@ -32,10 +32,10 @@ Export-Package: org.eclipse.glsp.ide.editor,
  org.eclipse.glsp.ide.editor.clipboard,
  org.eclipse.glsp.ide.editor.clipboard.ui,
  org.eclipse.glsp.ide.editor.di,
+ org.eclipse.glsp.ide.editor.gmodel.operations,
  org.eclipse.glsp.ide.editor.handlers,
  org.eclipse.glsp.ide.editor.initialization,
  org.eclipse.glsp.ide.editor.internal.actions,
- org.eclipse.glsp.ide.editor.operations.handlers,
  org.eclipse.glsp.ide.editor.ui,
  org.eclipse.glsp.ide.editor.utils
 Import-Package: javax.servlet;version="[3.1.0,4.0.0)",
